translate(custom): rename CUSTOM to menu_controls
