--- CUSTOM.H
@@ -1,6 +1,0 @@
-#ifndef CUSTOM_H
-#define CUSTOM_H
-
-void menu_customize_controls();
-
-#endif
