+++ menu_controls.h
@@ -1,0 +1,6 @@
+#ifndef MENU_CONTROLS_H
+#define MENU_CONTROLS_H
+
+void menu_customize_controls();
+
+#endif
